WIP on second_order_adjacent Elem functions...
Idea is to extend:
* n_second_order_adjacent_vertices()
* second_order_adjacent_vertex()
* _second_order_adjacent_vertices

which almost contain the information we need for
MeshTools::find_nodal_neighbors() to work correctly.

An issue seems to be that mid-face nodes (Quad9) are currently
considered to be second-order-adjacent to *all* of the mid-edge nodes
of the element, even though they are not connected by edges.  Likewise
the center node of a Hex27 is considered to be second-order-adjacent
to the 8 surrounding mid-edge nodes, but none of the 6 surrounding
mid-face nodes.  Whereas I would consider it to be a nodal neighbor of
no nodes.  Perhaps this is different enough to require a separate
interface...

Places where n_second_order_adjacent_vertices() is currently called are
* MeshTools::Modification::smooth()
* MeshTools::Modification::all_second_order()
* LaplaceMeshSmoother::smooth()

where in the latter case, they are used to move the central node into
the average position of the second-order-adjacent nodes during mesh
smoothing.