PointLocatorTree: avoid Elem::contains_point() warnings when using custom tolerances
We do this by actually calling Elem::close_to_point() instead of
Elem::contains_point() when the user sets a custom contains_point()
tolerance on the PointLocator.

Note that there are also contains_point() checks in
TreeNode<N>::find_element() that potentially use a custom tolerance,
so we should make the same change there as well.