Big cleanup.
Remove several "fallback" search methods and overall greatly simplify
the code. The basic idea is that in the parallel case, we expect to
have many "failed" searches, since the Point searched for will in most
cases live on a different processor. Therefore, we need to have failed
searches fail "quickly" and not spend lots of time exhaustively
searching. In order to balance between search robustness and speed,
the user can tune the _num_results parameter for their specific
application.