Update comment
We previously used the first non-zero Elem::hmin() value encountered
in the entire Mesh to set a "preliminary" h_min value, but this
approach had a couple drawbacks: (i) it could set and h_min value from
a part of the mesh not relevant to the part actually being stitched,
and (ii) it was not sufficient in the case where the entire mesh was
comprised of NodeElems

The approach that we use now of computing the minimum node separation
distance amongst all NodeElems on the boundary should give a much more
consistent value for h_min. In the case where the boundary being
stitched consists of a single NodeElem (hence there is no minimum node
separation or valid h_min value) we will still fall back on using an
absolute tolerance check as we did previously.